refactor: おすすめリスト行の旧形式フォールバックを撤去し、生成経路を単一の行形式に一本化
旧 .dat 形式への後方互換フォールバック（テンプレートの description 再計算・
関連タグの全体マップ読み）は、無くても移行期の表示劣化が軽微（説明文と
関連タグチップが最大1時間空になるだけ）な一方、消されずに残り続ける類の
コードなので撤去する。

- 行の整形を RecommendRowFormat に集約し、.dat 生成(bulk)とライブ生成
  (per-tag)の両経路が同じ最小10フィールド+desc40 を出力する
  （これまでライブ生成だけ旧形式の全フィールド行を返していた）
- テンプレートは desc40 を読むだけの単一経路に
- 関連タグは .dat 同梱値のみ参照し、未同梱は空表示（次の毎時生成で自然回復）。
  これに伴い不要になった StaticDataFile::getRelatedTags() を削除

移行期(デプロイ後〜次の毎時バッチ)の挙動: リスト説明文は空・関連タグチップは
汎用埋めになるが、ページは全て200で例外も出ないことを旧形式 .dat で確認済み。

検証 (本番同等データ):
- 新形式 .dat で対象10ページの修正前後HTMLバイト一致
- ライブ生成経路(.dat未生成タグ)も /oc はバイト一致、/recommend は
  関連タグチップのみ既定埋めに変わることを差分で確認
- 旧形式 .dat: 全ページ200・説明文のみ空・exception.log 0件
- 応答時間に変化なし (なりきり 10.3ms / oc192 8.3ms)
