+++ app/Services/Recommend/RecommendRankingBuilder.php
@@ -30,19 +30,26 @@ class RecommendRankingBuilder
         $limit = AppConfig::LIST_LIMIT_RECOMMEND;
 
         // 24時間の人数増加が大きい順（＝「いま伸びている」部屋）。
-        $growing = $repository->getRanking(
-            $entity,
-            AppConfig::RANKING_DAY_TABLE_NAME,
-            AppConfig::RECOMMEND_MIN_MEMBER_DIFF_H24,
-            $limit
+        // 行は .dat 生成(bulk)と同じ最小形式(RecommendRowFormat)に揃える。
+        $growing = array_map(
+            fn(array $row) => RecommendRowFormat::slim($row, $row['table_name'], (int)$row['diff_member_24h']),
+            $repository->getRanking(
+                $entity,
+                AppConfig::RANKING_DAY_TABLE_NAME,
+                AppConfig::RECOMMEND_MIN_MEMBER_DIFF_H24,
+                $limit
+            )
         );
 
         // 伸びていない部屋は member 降順で裾を埋める（痩せタグ対策・既存の大型部屋）。
         // 表示は30件のままだが、/oc 関連ルームの人数絞り込み母集団として300件保持する。
-        $member = $repository->getListOrderByMemberDesc(
-            $entity,
-            array_column($growing, 'id'),
-            AppConfig::LIST_LIMIT_RECOMMEND_POOL
+        $member = array_map(
+            fn(array $row) => RecommendRowFormat::slim($row, $row['table_name'], null),
+            $repository->getListOrderByMemberDesc(
+                $entity,
+                array_column($growing, 'id'),
+                AppConfig::LIST_LIMIT_RECOMMEND_POOL
+            )
         );
 
         // DTO は先頭(=表示順)に伸び部屋、末尾に裾を渡す（旧 hour/day/week の4段は廃止）。
