+++ app/Services/StaticData/StaticDataFile.php
@@ -84,21 +84,4 @@ class StaticDataFile
         return $data;
     }
 
-    /**
-     * 関連タグマップ（タグ => [関連タグ => 共起スコア]）。RelatedTagsService 参照。
-     *
-     * @return array<string, array<string, int>>
-     */
-    function getRelatedTags(): array
-    {
-        /** @var array|false $data */
-        $data = $this->fileStorage->getSerializedFile('@relatedTags');
-        if (!$data || AppConfig::$disableStaticDataFile) {
-            /** @var StaticDataGenerator $staticDataGenerator */
-            $staticDataGenerator = app(StaticDataGenerator::class);
-            $data = $staticDataGenerator->getRelatedTags();
-        }
-
-        return is_array($data) ? $data : [];
-    }
 }
