+++ app/Views/components/open_chat_list_recommend.php
@@ -2,7 +2,7 @@
   <?php /** @var \App\Services\Recommend\Dto\RecommendListDto $recommend */
 
   use App\Config\AppConfig;
-  use App\Views\Classes\CollapseKeywordEnumerations;
+  
 
   if (!isset($listArray)) {
     $listArray = $recommend->getList($shuffle ?? false, ($limit ?? null) ? AppConfig::$listLimitTopRanking : null, $id ?? 0);
@@ -28,20 +28,9 @@
           <a class="openchat-item-title unset" href="<?php echo url('/oc/' . $oc['id']) . ((isset($oc['table_name']) && ($oc['table_name'] === AppConfig::RANKING_HOUR_TABLE_NAME || $oc['table_name'] === AppConfig::RANKING_DAY_TABLE_NAME)) ? '?limit=hour' : '') ?>"><?php if (($oc['emblem'] ?? 0) === 1) : ?><span class="super-icon sp"></span><?php elseif (($oc['emblem'] ?? 0) === 2) : ?><span class="super-icon official"></span><?php endif ?><?php if (($oc['join_method_type'] ?? 0) === 2) : ?><span class="lock-icon"></span><?php endif ?><?php echo $oc['name'] ?></a>
         </h3>
         <p class="openchat-item-desc unset"><?php
-                                            // SEO: LINE 公式 description のコピー量を削減するため 40 字で truncate
-                                            // desc40 = .dat 生成時に collapse + truncate 済みの表示用説明文
-                                            // (BulkRecommendRankingBuilder::buildDisplayDescription)。
-                                            // 無い行(ライブ生成・旧形式 .dat)は従来どおり描画時に計算する。
-                                            // htmlspecialchars_decode はフレームワークの自動エスケープ経由(タグページ)と
-                                            // 生値経由(/ocのviewComponent)の両方を原文に正規化するためのもの。
-                                            if (isset($oc['desc40'])) {
-                                              $displayDesc = htmlspecialchars_decode($oc['desc40']);
-                                            } else {
-                                              $collapsedDesc = CollapseKeywordEnumerations::collapse(htmlspecialchars_decode($oc['description'] ?? ''), extraText: htmlspecialchars_decode($oc['name']));
-                                              $displayDesc = mb_strlen($collapsedDesc) > 40 ? mb_substr($collapsedDesc, 0, 40) . '…' : $collapsedDesc;
-                                            }
-                                            echo h($displayDesc);
-                                            ?></p>
+                                            // 説明文は行の生成時に collapse + 40字 truncate 済み (RecommendRowFormat)。
+                                            // decode は自動エスケープ経由(タグページ)と生値経由(/oc)の両方を原文へ正規化する。
+                                            echo h(htmlspecialchars_decode($oc['desc40'] ?? '')) ?></p>
         <footer class="openchat-item-lower-outer">
           <div class="openchat-item-lower unset" style="font-size: 13px; margin-top: 0;">
             <?php if (isset($oc['member'])) : ?>
